requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==3.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.4
dnspython==2.3.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
PyJWT==2.7.0
pymongo==3.12.1
pytz==2023.3
sqlparse==0.2.4
tzdata==2023.3
